+++ src/Zilf/Interpreter/FileFlags.cs
@@ -26,6 +26,8 @@ namespace Zilf.Interpreter
         None = 0,
         CleanStack = 1,
         MdlZil = 2,
-        SentenceEnds = 3,
+        SentenceEnds = 4,
+        KeepRoutines = 8,
+        SuppressUnusedRoutineWarnings = 16,
     }
 }